create AuthForm component
